Use func.func and func.return in the spec
The stablehlo.func and stablehlo.return syntax in the opening example
of the spec was aspirational, and at the time we expected that we'll
soon adopt it.

A lot of the time has passed, and we still haven't gotten around to
doing this, while we've been getting regular questions about this.
Let's reflect reality in the example for now - we can always change
it back once we address #425.