Tighten the formal notation that we're using in the spec
This PR proposes a more consistent notation for spec.md and describes
it in the newly introduced "Notation" section. It doesn't address all
feedback from #289, and some areas are deliberately left for future work
but I think it's already an improvement.

1) Introduces the ellipsis notation, to replace previously inconsistent
use of colons in a similar role, to allow writing things like:
`dim(lhs, lhs_batching_dimensions...) = dim(rhs, rhs_batching_dimensions...)`.

As a side effect, this also gets rid of most of the `i0`, `i1`, etc,
except in two places: a) when we mix scalar indices with colons, e.g.
in GatherOp or ScatterOp, b) when we create function types from
individual input/output types. In these situations, I believe that
scalar indices represent the best tradeoff between conciseness and
simplicity of notation.

Rationale: A lot of formulas become a lot more concise thanks to
the ellipsis notation, with most of "`<formula with some scalar>` for
all `scalar` in `something`" formulas replaced with just
`<formula with ellipses>`.

2) Consistently switches to using programming syntax rather than LaTeX
syntax, as discussed offline.

Rationale: LaTeX syntax looks pretty, but it is not copypasteable when
rendered on GitHub (huge problem) and is harder to read and write in
Markdown (considerable problem).

3) Consistently switches to the "`size(foo)`" rather than alternatives
along the lines of "size(`foo`)".

Rationale: This makes it considerably easier to read and write Markdown
representation of formulas because fewer special symbols are involved.

4) Prefers formulas like "`type(foo) = type(bar)`" to prose like
"`foo` and `bar` have the same type" wherever practical. There are just
a handful of exceptions where formulas would be very heaviweight.

Rationale: This is the direction that the spec has already been leaning
towards, I just helped that a little bit.

5) Adds some new non-terminals and renames some existing ones to make
sure that functions used in formulas derive cleanly from them.

Rationale: I think that we need blanket rules along the lines of
"if type's non-terminal is called FooType, then there is automatically
a function called is_foo which is available for types and values".
To make that happen, I tweaked the names a little bit.

6) Retires the "`foo in [lo, hi)`" syntax in favor of the
`lo <= foo < hi` syntax.

Rationale: The former syntax only works for scalars, the latter syntax
consistently applies both to scalars and tensors (via implicit
broadcasting).

7) Consistently switches all uses of `==` to use `=`.

Rationale: This is the direction that the spec has already been leaning
towards, I just helped that a little bit. Even though we use `=` both
for comparisons and for assignment, these usages can be unambiguously
discerned, so I propose we keep `=` since it's more concise.

8) Introduces uniform syntax for an equivalent of switch expressions:
"`foo` is defined as: * `value1` if `condition1 * `value2` otherwise".

Rationale: There are several ways of writing this in the current version
of the spec, so I figured it would be useful to unify.

9) (This is a controversial one, and maybe I'm wrong here). Switches
most of formulas like `foo > lo` to say `lo < foo` to be consistent with
many formulas like `lo < bar < hi`.