Fail interpretation on Not in HLO and Miscellaneous ops
As we're getting closer and closer to completion of the intepreter
thanks to great work by Gunhyun, our corner cases are becoming more
and more prominent.

One of these corner cases is support for Not in HLO ops, which we
are planning to remove from the StableHLO opset but which are still
part of the StableHLO dialect.

Another corner case is support for Miscellaneous ops, all of which
except for FftOp have decomposition passes implemented either on MHLO
or HLO. Some of these decompositions are really tricky, and it would
be a shame to duplicate their code and to maintain this duplicated
code in the interpreter.

As a result, I propose that we fail when encountering these ops
and leave decomposition of these ops to be the responsibility of
the users of the interpreter for now.

In the future, we should provide for these ops in this repository
(or move these ops out of the StableHLO opset). We can do this via a
a combination of moving the already existing MLIR-based passes to
this repository (and running them before the interpreter runs) and/or
porting the already existing HLO-based passes to MLIR in this
repository.

This proposal also involves deleting the already implemented eval
functions and tests for BatchNormGradOp, BatchNormInferenceOp,
BatchNormTrainingOp and CholeskyOp. A lot of work went into these
implementations (which will not be forgotten!), and I'm hesitating to
remove them. However, this is a non-trivial amount of code and it may
need non-trivial amount of maintenance down the line, so my
recommendation would be to remove it.