Update spec and interpreter for ReduceWindowOp to properly calculate window_end (#1492)
The current spec of ReduceWindowOp does not use `slice` with the same semantics 
as what `stablehlo.slice` does. `stablehlo.slice` version of `limit_indices`
accounts for dilation and also complies with SliceOp's constraint: `(C3)
0 <= start_indices[d] <= limit_indices[d] <= dim(operand, d) for all
dimension d.`

ReduceWindowOp's `limit_indices` calculation does not account for the
dilation and (C3), so we need to add dilation (minus 1 since
limit_indices is exclusive).

This was discussed previously in the interpreter PR
https://github.com/openxla/stablehlo/pull/1336#discussion_r1142826672.
We kept the formulation as-is to assume that the program provides
correct values, but we may not assume this is the case when we are
explicitly constructing the input values.