Update compatibility.md (#1606)
We have recently approved #1600 - an RFC about compatibility. This PR
updates compatibility.md - the source of truth for StableHLO's
compatibility guarantees - to the latest plan of record.

Also, this PR rolls back changes to the original compatibility RFC done
in #1306. I now think that it was a mistake to retroactively update it,
even if the plan of record changed - it is much easier to reason about
RFCs if they are immutable.