fix bug in verification in ReduceOp (#1613)
It seems that there is a bug in the verifier of `reduce` op when all the
inputs are un-ranked. Specifically,

```
if ((!allInputsUnranked &&
         dimension >= inputTypes[rankedInputIdx].getRank()) ||
        dimension < 0)
      return emitOptionalError(
          location, "Out-of-bounds dimension ", dimension,
          " for input-tensor rank: ", inputTypes[rankedInputIdx].getRank());
```

If `allInputsUnranked` is true and `dimension < 0`, then while
evaluating the error message ` inputTypes[rankedInputIdx].getRank()` we
will incur an invalid memory access as `rankedInputIdx = -1` when
`allInputsUnranked = true`.

Added a test to reproduce that.