Specification for quantized ConvolutionOp (#1477)
## Summary 
The PR proposes the specification for quantized convolution op along
with the specifications for a few other ops on which the convolution
specification depends on, for example, `pad`, `reverse`.

## A few details
1. Given `q = tensor with uniformed quantized type`, the PR covers the
semantics of static range quantized `convolution(q, q)`. The
specification is very similar to the earlier proposed [quantized
dot_general
op](https://github.com/openxla/stablehlo/blob/main/docs/spec.md#dot_general).

2. Given, that many ops support both per-tensor and per-axis
quantization schemes, it might be redundant to specify for each op the
semantics of the computation involved in both the scheme. To avoid that
we decided that the op-level semantics will only specify the per-tensor
semantics, whereas, semantics of per-axis computations will be described
generally in one place.