Avoid out of bound reads in the spec of GatherOp (#1541)
In a recent
[discussion](https://github.com/openxla/stablehlo/pull/1058#discussion_r1203386357)
on the GatherOp implementation, we pointed out that the semantics did
not fully address what to do with out of bounds indexes. To make sure
the bounds are always correct, clamping on the `start_index` is needed.
This reflects the change needed in the spec and the reference
implementation.