[RFC] StableHLO Extensibility
This PR does not contain an actual RFC - it's merely tracking a
discussion on openxla-discuss: https://groups.google.com/a/openxla.org/g/openxla-discuss/c/Ao5K8fvXoEk.
Once this RFC reaches an outcome, we'll update the status field in this
markdown file and merge this PR to document that.