Add constraints to some quantized ops related to per-axis scheme (#1535)
### Summary 

The PR augments the a few ops, whose specs is already published, with
constraints in the following way:

Constraints added w.r.t the per-axis qunatization scheme,
 - **transpose**:  Transpose the `quantization_dimension` of `output`.
- **reshape**: Avoid merging or splitting of elements along
`quantization dimension` of `operand`.
 - **slice**: Slice the `zero_points(operand)` and `scales(operand)`.
- **reverse**: reverse the `scales` and `zero_points` of the `result` if
the `quantization_dimension(operand)` is in `dimensions` to reverse.
- **pad**: The op is simplified to use only per-tensor scheme. The
rationale for this is as follows:
- If the `padding_value` is added in axis other than
`quantization_dimension(operand)`, then the `zero_point(padding_value)`
should match the `zero_points(operand)[i]` for all `i`. That means all
the values in `zero_points(operand)` are the same, which is basically
the per-tensor case.
- Only meaningful case is if the `padding_value` is added __only__ along
the `quantization_dimension(operand)`. In this case, the
`zero_points(result)` and `scales(result)` need to be updated
accordingly. I am not if there is a use case for this particular case
and hence kept things simple by allowing on per-tensor scheme for this
op.

Constraints added to comply with other similar specs
- **add**: For `QuantizedTensorType`, we introduced a new non-terminal
`QuantizedElementType` which is different from the non-terminal
`ElementType` used for `TensorType`. The proposed change accommodates
that.

Please let me know your feedback. 

PS: The PR involves shuffling the constraint labels (C*). Once the PR
reaches a consensus, I will update the verification/type infenrence code
to match the constraint labels.

**update**: As per my
[proposal](https://groups.google.com/a/openxla.org/g/openxla-discuss/c/iwE9is49SS4/m/mi9AyvKZAQAJ),
I have updated the PR to specify the semantics using per-tensor scheme.
We can find that it greatly simplifies the specification. With that my
proposal is to add per-axis details on a need basis.

**update 06/20/2023**

It turns out that the current PRs becomes "empty" after a few PRs are
merged after the introduction of the current PR.

I note that:
1. Some of the ops addressed in the current PRs (like `transpose, slice,
reshape`) have already been introduced with quantized type in #1413
which got merged after the PR was introduced.
2. Other ops addressed in the current PRs (like `pad, reverse`) are also
been introduced with quantized type in #1477 which got merged after the
PR was introduced.
3. The newly introduced constraints in the current PRs are all
simplified while rebasing with
https://github.com/openxla/stablehlo/pull/1629.