Clarify what "unspecced features" means
Based on a conversation with Kevin from earlier today. This is a nice
catch and an important clarification.

We do want to provide compatibility guarantees around new features
(e.g. make sure that they cannot be serialized to portable artifacts
with older target versions) - unspecced features mean something
different, which is explicitly elaborated in the document now.