Offer StableHLO portable artifact deserialization as a pass pipeline (#1651)
Some users want to write APIs that accept portable artifacts _and_ other
programs. Currently adding support for portable artifacts requires
additional logic to test if input is a portable artifact, then
deserialize it before running any other passes. This is not ergonomic,
as the more MLIR way of doing this is to offer a pass that will
deserialize _if_ the input is a portable artifact - for example, support
for StableHLO was added to some TF APIs using a
`createHloLegalizeToStablehlo` pass.

To get to a point to permit this, we need two changes:

1. Don't error on non-VHLO dialects when deserializing, currently only
happens in `--vhlo-to-version`.
This isn't a big issue, as `--stablehlo-legalize-to-vhlo` already
validates that programs are stable. The current way VHLO to version is
written is already more lenient than StableHLO->VHLO, as it only errors
on Func and StableHLO ops, meaning, for example, that arith ops are
permitted even without this change (would be caught by StableHLO->VHLO
though).

2. Add deserialization pipeline. This allows users of StableHLO to add
these passes to their input pipeline to quickly support StableHLO
portable artifacts.

Part of https://github.com/openxla/stablehlo/issues/1530.