[RFC] Forward Compatibility Testing (#1498)
This is a proposal for forward compatibility testing in StableHLO. This
ensures that payloads serialized at HEAD can be deserialized at a
previous release.

Note, this plan is built on top of our existing backward compatibility
testing infrastructure, and as such does not include semantic forward
compatibility testing. Once we have addressed semantic compatibility
testing in #1416 we can expand this machinery to include semantic
forward compatibility testing.

- **Proposed design prototype:**
https://github.com/GleasonK/stablehlo/pull/35
- **Alternate design prototype:**
https://github.com/GleasonK/stablehlo/pull/33

Notification of this PR will be sent to the
[openxla-discuss](https://groups.google.com/a/openxla.org/g/openxla-discuss)
group.

Related to #1318.

UPDATE 5/25: While this RFC is in review, some form of forward
incompatibility detection is still required by StableHLO. In
https://github.com/openxla/stablehlo/pull/1525 a stop-gap solution based
on the preferred design was implemented. It will be modified in
accordance with feedback on this RFC.