Specification of quantized ops from `Control-flow`, `Distribution`, `Extensibility`, and `Miscellaneous` categories.  (#1647)
### Summary 

This PRs provides the specification of the ops from the following
categories:

* **Control-flow ops**: `after_all, case, if, optimization_barrier,
while`
- `afer_all`: Not applicable for quantization as the ops deals with
setting data-flow ordering using tokens.
- `case, if`: These ops implicitly captures operands from the parent
scope. The result of these operations and the corresponding blocks can
be quantized types.
- `optimization_barrier`: The `operand` and `result` can both be
quantized tensors.
- `while`: The `operand` and `result` can both be quantized tensors and
of same type. Per the semantics of `while` op, the output of the
associated block is used for the next iteration with the `operand` used
for the very first iteration. With that, I do not think we can use the
`operand` and `result` with different quantization types.
 
* **Distribution**: `all_gather, all_reduce, all_to_all,
collective_permute, infeed, outfeed, partition_id, recv, reduce_scatter,
replica_id, send`
All the ops in this category except for the following ones will support
quantized types:
- `all_reduce, reduce_scatter`: These involve quantization of reduction
operation https://github.com/openxla/stablehlo/pull/1538. I would
propose to address these two after the quantization of reduction ops are
addressed.
  - `partition_id, replica_id`: Not applicable 

* **Extensibility**: `custom_call, get_tuple_element, tuple`
All the ops in this category supports all the `ValueType`s including
`QuantizedTensorType`. No change is needed in the existing spec.

* **Miscellaneous**: `batch_norm_grad, batch_norm_inference,
batch_norm_training, cholesky, constant, fft, iota, rng,
rng_bit_generator, triangular_solve`
- The specification of `batch_norm_inference`, `cholesky`,
`triangular_solve` ops follow the same `dequantize-op-quantize` strategy
as we followed for the elementwise ops.
- `batch_norm_grad, batch_norm_training`: These involve quantization of
reduction operation https://github.com/openxla/stablehlo/pull/1538. I
would propose to address these two after the quantization of reduction
ops are addressed.
- `constant, iota`: The output type of these ops can be a quantized type
of user-defined quantization parameters.
- `fft`: This op supports `element_type(operand) = element_type(result)
= complex type`, when the `fft_type` is `FFT` or `IFFT`. For `fft_type =
RFFT`, the op supports floating-point `operand` and complex type
`result` and for `fft_type = IRFFT`, it supports complex type `operand`
and floating-point `result`. IMO, only for `fft_type = RFFT`, the
operand can support quantized type and the semantics could be to
quantize the operand to floating-point before doing the computation.
- `rng`: The `operand` and `result` can be of quantized type only when
the `rng_distribution` is `NORMAL` in which case the semantics follows
the `dequantize_op_quantize` strategy. The motivation for this choice
comes from the fact the op supports floating point type only when
`rng_distribution` is `NORMAL` and we propose to treat this op, on
floating-point tensors, as an op on quantized tensor, along the lines of
dequantize -> float computation -> quantize.
- `rng_bit_generator`: To be discussed in "open questions" section
below.

### Open questions
- `rng_bit_generator`: This op generates a randomized output tensor
based on a `rng_algorithm` algorithm (`DEFAULT`, `THREE_FRY`, and
`PHILOX`). It is not clear how there algorithm should handle generating
quantized output. My proposal here is to skip quantization of this op
unless we have a use case for it. Please let me know your opinion on
this.

Please review and let me know your feedback. 



**update 06/26/2023**
Scoping `rng` out from quantization based on
https://github.com/openxla/stablehlo/pull/1647#issuecomment-1607857449

**update 07/13/2023**
Per
https://github.com/openxla/stablehlo/pull/1647#discussion_r1251233481, I
proposed not to fuse de-quantization of operand and quantization of
result for `fft_type = RFFT` and `fft_type = IRFFT` respectively for fft
op. With that a user can explicit de-quantize of operand or quantize of
result as needed.

In the current proposal, the `batch_norm_inference` op, do not consider
the constant operand `epsilon` to be of quantized type. IMO, considering
that operand with quantized type unnecessary complicate the
expressiveness of the op.