Add Interpreter dialect (#1703)
The StableHLO reference implementation is nearing its final phase to
implement most of the specced operations in the StableHLO opset, and one
major remaining piece is the Distribution ops. This PR creates the
`Interpreter` dialect specifically targeted for interpreter-related ops
that enable running parallel execution of Distribution ops. In the
future, this dialect may also be used to add additional functionality to
the interpreter.