Using dequant-op-quant for convolution and dot_general (#1700)
fixes #1628 

The idea behind the PR is to treat the quantization of `dot_general` and
`convolution` similar to how various other element wise ops are
specified. Other than the obvious parity of the specification with other
ops, there are other benefits too.

1. `dot_general` and `convolution` are potential candidates for per-axis
quantization. With this solution we can define per-axis scheme
uniformaly in the meta-functions `dequantize` and `quantize` and thereby
do not have to explicitly handle the per-axis scheme for these ops.
2. `dot_general` and `convolution` uses some data_movement ops (like
`reshape`, `transpose`, `pad`, `slice`) for the underlying
specification. There is a concern where the fact that `dot_general` and
`convolution` can potentially use per-axis quantization necessitates
handling of the per-axis scheme for the data movement ops just because
the later are used in the specification of the former. With the current
solution, we wouldn't need to worry about per-axis quantization for the
underlying data movement ops.