Add interpreter for OutfeedOp (#1723)
We have the following constraints in the spec:

```
(I1) `inputs`: is a variadic number of tensors.
(I2) `token`: is a `token`.
(I3) `outfeed_config`: is a constant of type `string`.
(C1) `type(result) = token`.
```

These constraints will be comprehensively covered by the following
tests:

```
I1: a) `inputs`: is not a variadic number of tensors. (Covered by ODS).
I2: a) `token`: is not a `token`. (Covered by ODS).
I3: a) `outfeed_config` is not a constant of type `string`. (Covered by ODS).
C1: a) `type(result) != token`.
```

If we drop the "Covered by ODS" pieces, this will leave us with the
following test cases:

```
C1: a) `type(result) != token`.
```

closes #1130