build: use 64-bit integer literals across all platforms (#1760)
The `long` type is 32 bits wide on Windows, either 32 or 64 bits wide on
Linux (depending on x86 or x64), and 64 bits wide on macOS.
Consequently, when we use bit-shift operators to construct 64-bit
integers or when we check for splat tensors, using (signed or unsigned)
`long` integers can produce incorrect values.

This patch replaces signed and unsigned `long` integer literals with
signed and unsigned `long long` values respectively, so that the
they are always 64 bits wide on all platforms.  As a minor change, this
patch uses the lower-case suffixes (`{u}ll` instead of `{U}LL`) for
consistency.

This patch is in anticipation of downstream users of StableHlo, like
Torch-MLIR (which use Windows and macOS builds), swapping MLIR-HLO with
StableHlo.