build: miscellaneous CMake fixes (#1766)
This patch makes two small changes.  First, it adds the `MLIRParser`
library as a dependency for the `StablehloPortableApi` and the
`StablehloSerialization` library targets, since both PortableApi.cpp and
Serialization.cpp include calls to `mlir::parseSourceString()`.
Although the code _does_ build without this dependency, it seems like an
accidental result, so this patch makes the dependency explicit.

Second, the Windows shell does not use the same rules as Bash for
setting environment variables, thus causing the `check-stablehlo-python`
target to fail.  This patch uses the (cross-platform) CMake syntax for
setting the `PYTHONPATH` environment variable before running the python
script.