Fixing the access of scalar elements in `broadcast_in_dim`, `clamp`, and `select` operations (#1777)
Semantics of some operations like `broadcast_in_dim`, `clamp`, and
`select`, with one or more of the input operands potentially be scalars,
needs implicit broadcasting. For example, `operand` for
`broadcast_in_dim`, `min`/`max` for `clamp`, `pred` for `select`.

The current PR fixes the access of those operands, to indicate proper
deference of the tensor type to element type, in the specification of
these operations.