Fixing verifier for reshape, transpose, and broadcast_in_dim ops (#1782)
fixes https://github.com/openxla/stablehlo/issues/1765

The existing verifier for these ops was using the trait
[SameOperandsAndResultElementType](https://github.com/openxla/stablehlo/blob/b2090e9774f68c81b65615d18872b99b718eedad/stablehlo/dialect/StablehloOps.td#L2763)
to check the equality of the element types which will not work for the
per-axis quantized variants of these ops as the quantization dimension
and quantization parameters (scales and zero points) may differ between
operand and result.

The PR proposed to replace the trait with
[HLO_CompatibleOperandsAndResultElementType
](https://source.corp.google.com/piper///depot/google3/third_party/stablehlo/stablehlo/dialect/Base.td;rcl=567468454;l=216)(which
just checks the equality for storage type and expressed type).

Note that the PR does not provide a comprehensive solution here. It just
allows the representational capability to these ops thereby allowing
writing meaningful quantized versions of the ops w/o verification error.
A comprehensive verification solution should include verifying all the
constraints related to quantization for these ops and is true for other
quantized ops as well. That would need additional work and will be
addressed collectively for the full set of quantized ops.