Add interpreter for SendOp and RecvOp (#1732)
SendOp:
We have the following constraints in the spec:

```
(I1) `inputs` is a variadic number of tensors.
(I2) `token` is a `token`.
(I3) `channel_id` is a constant of type `si64`.
(I4) `channel_type` is a enum of `DEVICE_TO_DEVICE` and `DEVICE_TO_HOST`.
(I5) `is_host_transfer` is a constant of type `i1`.
(C1) `channel_type` is defined as:
* `DEVICE_TO_HOST` if `is_host_transfer = true`,
* `DEVICE_TO_DEVICE` otherwise.
```

These constraints will be comprehensively covered by the following
tests:

```
I1: a) `inputs` is not a variadic number of tensors. (Covered by ODS).
I2: a) `token` is not a `token`. (Covered by ODS).
I3: a) `channel_id` is not a constant of type `si64`. (Covered by ODS).
I4: a) `channel_type` is not a enum of `DEVICE_TO_DEVICE` and `DEVICE_TO_HOST`.
I5: a) `is_host_transfer` is not a constant of type `i1`. (Covered by ODS).
C1: a) `channel_type` is not defined as:
* `DEVICE_TO_HOST` if `is_host_transfer = true`,
* `DEVICE_TO_DEVICE` otherwise.
```

If we drop the "Covered by ODS" pieces, this will leave us with the
following test cases:

```
I4: a) `channel_type` is not a enum of `DEVICE_TO_DEVICE` and `DEVICE_TO_HOST`.
C1: a) `channel_type` is not defined as:
* `DEVICE_TO_HOST` if `is_host_transfer = true`,
* `DEVICE_TO_DEVICE` otherwise.
```

closes #1137

RecvOp:
We have the following constraints in the spec:

```
(I1) `token` is a `token`.
(I2) `channel_id` is a constant of type `si64`.
(I3) `channel_type` is a enum of `DEVICE_TO_DEVICE` and `HOST_TO_DEVICE`.
(I4) `is_host_transfer` is a constant of type `i1`.
(C1) `channel_type` is defined as:
* `HOST_TO_DEVICE` if `is_host_transfer = true`,
* `DEVICE_TO_DEVICE` otherwise.
(C2) `0 < size(results)`.
(C3) `is_empty(result[:-1])` or `is_tensor(type(results[:-1]))`.
(C4) `is_token(type(results[-1]))`.
```

These constraints will be comprehensively covered by the following
tests:

```
I1: a) `token` is not a `token`. (Covered by ODS).
I2: a) `channel_id` is not a constant of type `si64`. (Covered by ODS).
I3: a) `channel_type` is not a enum of `DEVICE_TO_DEVICE` and `HOST_TO_DEVICE`.
I4: a) `is_host_transfer` is not a constant of type `i1`. (Covered by ODS).
C1: a) `channel_type` is not defined as:
* `HOST_TO_DEVICE` if `is_host_transfer = true`,
* `DEVICE_TO_DEVICE` otherwise.
C2: a) `size(results) <= 0`.
C3: a) `is_empty(result[:-1]) = false` and `is_tensor(type(results[:-1])) = false`.
C4: a) `is_token(type(results[-1])) = false`.
```

If we drop the "Covered by ODS" pieces, this will leave us with the
following test cases:

```
I3: a) `channel_type` is not a enum of `DEVICE_TO_DEVICE` and `HOST_TO_DEVICE`.
C1: a) `channel_type` is not defined as:
* `HOST_TO_DEVICE` if `is_host_transfer = true`,
* `DEVICE_TO_DEVICE` otherwise.
C2: a) `size(results) <= 0`.
C3: a) `is_empty(result[:-1]) = false` and `is_tensor(type(results[:-1])) = false`.
C4: a) `is_token(type(results[-1])) = false`.
```

Notes:
* Also fixes bugs in the compatibility tests. The changes made here is
an exception to the regular process (See
[compatibility.md](https://github.com/openxla/stablehlo/blob/main/docs/compatibility.md#out-of-scope))
since the verification was not in sync with the spec. Other than the bug
fixes, the remaining changes were autogenerated using the script in
[compatibility.md](https://github.com/openxla/stablehlo/blob/main/docs/compatibility.md).

closes #1132