Add reference interpreter probe instrumentation (#1784)
Reopening #1763 due to a branch corruption issue. Original text below:

===

As the number of ML frameworks continues to rapidly grow, it’s becoming
increasingly desirable to be able to compare numerical results across
different runtimes. This ensures that ML models run equivalently in
different environments, and overall helps to instill confidence in ML
frameworks. For example, given the recency of the StableHLO
representation, it may be desirable to have a way to compare StableHLO
interpreter results to other frameworks, such as Tensorflow, or other
implementations

This PR proposes the addition of a probe instruction to the new
interpreter dialect in StableHLO. The probe instruction will enable us
to extract intermediate tensor state from the StableHLO interpreter.
Since this new op is intended to initially support debuggability, it
will not be added to the StableHLO opset itself, but instead to the new
[interpreter dialect](https://github.com/openxla/stablehlo/pull/1703).

The probe instruction will take precisely 1 shaped tensor input, and
produce the same shaped tensor output, without modifying it:

```
%1 = "interpreter.probe"(%0) {probe_id="probe0"} : (tensor<*xf32) -> tensor<*xf32>
```

The op will be modeled as having side effects, to prevent it from being
moved or removed by the compiler. It will additionally contain a
singular, unique string attribute, `probe_id`, which will allow us to
later pair this tensor’s data with other runtimes. At compile time, the
probe instruction may be inserted after every StableHLO operation
(except constants), allowing runtime data to be serialized to disk.

This PR also adds a new StableHLO interpreter configuration flag,
`--probe_readout_dir`, which will be used to specify the directory in
which to store all probed interpreter values. When executed, the probe
instruction will associate the contents of its input tensor with the
given probe_id, and serialize the tensor data to disk using the portable
[NumPy](https://numpy.org/) format. The filename used will be in the
form of `probe_readout_dir/probe_id.npy`. As the interpreter executes
`interpreter.probe` operations, in addition to serializing tensor data
to NumPy files, it will also append serialized filepaths to a metadata
`index.csv` file in the form of:

```
probe_id0,/probe_readout_dir/probe_id0.npy
probe_id1,/probe_readout_dir/probe_id1.npy
…
probe_id0,/probe_readout_dir/probe_id0_1.npy
```

The metadata file can then be used by post processing tools to analyze
the instrumented tensor data. Note that a `probe_id` can appear more
than once in the metadata file (i.e. if a probe is executed within a
loop).

### Validation

To validate the new instruction, we propose extending the existing
`stablehlo::check` MLIR dialect with a `expect_serialized_eq` operation,
which can be utilized in the following way:

```
// Where `T` is a statically shaped tensor
%0 = …
%1 = interpreter.probe %0 { probe_id = "probe0" } : (tensor<T>) -> tensor<T>
check.expect_serialized_eq %3 { probe_id = "probe0", iteration = 0 : i32 } : tensor<T>
```

The `check.expect_serialized_eq` is expected to appear after the
producing `interpreter.probe` instruction with the same `probe_id`. When
executed, the instruction will search for an existing `index.csv`
metadata file in the same directory as specified by the
`--probe_readout_dir` flag, and then attempt to find the serialized
filepath for the given `probe_id` and iteration. After which, the
interpreter will load the serialized tensor from the NumPy file, and
then proceed to compare it with its operand tensor. The operation will
assert on an answer mismatch or failure to find/parse the associated
NumPy file. The iteration attribute is 0-based and is optional and will
default to 0.