Introduce StableHLO interpreter API (#1797)
Today, it isn't practical to call the StableHLO reference interpreter.
Invoking the interpreter requires going through the MLIR translate
infrastructure (i.e. via `interpretRegistration` in
`StablehloTranslateMain.cpp`), and does not currently support specifying
user inputs or doing any meaningful post processing on the model outputs
short of printing them to screen. This makes it difficult to build
tooling which uses the reference interpreter.

This PR proposes a new reference interpreter API which facilitates the
running the StableHLO reference interpreter through a simple, unified
interface. At a high level, the proposed interface allows clients to
specify a set of inputs (in `InterpreterValue` format), and an
`InterpreterConfiguration` object (see below). The proposed interface
invokes the reference interpreter and returns a set of outputs (in
`InterpreterValue` format). The new interface would live in
`stablehlo/reference/InterpreterApi.h` and would expose the following
interface:

```
/// Invoke the StableHLO reference interpreter with the given parsed MLIR
/// module input and provided inputs. Returns a list of interpreter outputs.
/// Can optionally pass a fallback interpreter callback which executes when no
/// builtin kernels are matched.
llvm::ErrorOr<SmallVector<InterpreterValue>> evalModule(
    ModuleOp module, ArrayRef<InterpreterValue> inputs,
    const InterpreterConfiguration &config);
```


### Interpreter configuration
As part of this PR, there is an `InterpreterConfiguration` struct which
can be used to configure the StableHLO interpreter. Encapsulating
configuration state into a singular object allows the interface to
remain stable as new configuration options are added. Initially, the
configuration object will contain a handful of options:

```
struct InterpreterConfiguration {
  InterpreterConfiguration()
      : fallback(std::make_unique<InterpreterFallback>()) {}

  /// If specified, the directory to which StableHLO interpreter tensors will
  /// be serialized to disk.
  std::string probeInstrumentationDir = "";

  /// Use the specified named function as the main entrypoint into a module.
  /// Defaults to `main` for modules with multiple functions. If a module only
  /// contains 1 function and the default `main` value is used, the singular
  /// function will be used as the entrypoint (irrespective of a function name
  /// match).
  std::string mainFunction = "main";

  /// If specified, use the callback to run on ops which do not have a
  /// registered kernel.
  std::unique_ptr<InterpreterFallback> fallback;
};
```

Note that all of these configuration values are optional, and it is
possible to run the interpreter with a default initialized
`InterpreterConfiguration` struct (which will by default, not provide a
fallback callback).

### Interpreter fallback callback
When defining an interpreter fallback callback, the
`InterpreterFallback` struct can be extended:

```
class InterpreterFallback {
 public:
  /// Custom op kernels for any user specified ops not found in the StableHLO
  /// op dialect or StableHLO interpreter dialect.
  virtual llvm::Error operator()(Operation &op, Scope &scope, Process *process);

  virtual ~InterpreterFallback() = default;
};
```
Note that the user specified fallback need not contain any
`InterpreterOp` kernels, these will always run before the user specified
callback.

Currently, the proposed change exposes `InterpreterConfiguration` and
the current `mlir::func::FuncOp` being evaluated to the fallback
interface to support the existing flow. Consequently, there’s a bit of
gymnastics in `stablehlo::runInterpreter` which I think could be
improved on. I'm open to ideas on how we could better handle this.