[build] create initial CMake export configuration
Prior to this patch, it seems that consumers of StableHLO dialects and
passes are required to build StableHLO from source.  More precisely,
such consumers cannot point to a StableHLO build tree or install tree.
This is because StableHLO's build and install trees do not create an
export configuration (StablehloConfig.cmake) that consumers can refer
to, by calling `find_package(StableHLO)`.

This patch adds a CMake script to create a basic export configuration
that lists the library targets of StableHLO, along with their
dependencies.

Since these library targets are created using the `add_mlir_library()`
CMake function, these targets get added to MLIRTargets and MLIR_EXPORTS.
A better alternative is to add the names of these targets to StableHLO's
list of exports that is separate from MLIR's, but doing so requires a
somewhat invasive change of replacing all `add_mlir_library()` with a
wrapper that adds the target names to a StableHLO-specific list of
exports.